hw4/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
cssselect==1.3.0
defusedxml==0.7.1
ecdsa==0.19.1
elastic-transport==8.17.1
elasticsearch==8.18.1
fastapi==0.115.12
filelock==3.18.0
flake8==7.2.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
jieba==0.42.1
jmespath==1.0.1
joblib==1.5.1
lxml==5.4.0
mccabe==0.7.0
multidict==6.4.4
mypy==1.16.0
mypy_extensions==1.1.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
parsel==1.10.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
redis==6.2.0
requests==2.32.4
requests-file==2.1.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
Scrapy==2.13.2
scrapy-redis==0.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
tldextract==5.3.0
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
w3lib==2.3.1
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
